+++ CHANGELOG.md
@@ -14,6 +14,11 @@ dev
 * Added a new option: `notify_on_filtered_messages`. By turning it `on`, you
   can instruct the plugin to send notifications also for filtered (hidden)
   messages.
+* Improved the detection of nicks from data passed by WeeChat. Originally, the
+  nick was obtained from the prefix. However, the prefix is not always the
+  nick. Now, the nick is first tried to be obtained from a tag of the form
+  `nick_XYZ`, where `XYZ` is the nick that sent the message. If this fails, the
+  nick is obtained from the prefix.
 * Improved the removal of modes from prefixes. Originally, only `@` (op on IRC)
   and `+` (voice on IRC) were removed. Now, any character from the following
   list is removed: `~&@%+-`. The meaning depends on the used protocol.